refactor: Extracted native classes to tests/helpers
