+++ test-app/tests/helpers/autotracking/index.ts
@@ -1,0 +1,4 @@
+export * from './currency';
+export * from './population';
+export * from './royalty';
+export * from './user';
