+++ test-app/tests/helpers/autotracking/user.ts
@@ -1,0 +1,13 @@
+import { tracked } from '@glimmer/tracking';
+
+export class User {
+  @tracked _name = 'Zoey';
+
+  get name(): string {
+    return this._name;
+  }
+
+  set name(name: string) {
+    this._name = name;
+  }
+}
