+++ test-app/tests/helpers/index.ts
@@ -41,3 +41,5 @@ function setupTest(hooks: NestedHooks, options?: SetupTestOptions) {
 }
 
 export { setupApplicationTest, setupRenderingTest, setupTest };
+
+export * from './autotracking';
